Backend/FlowBee/pyproject.toml
[tool.poetry]
name = "flowbee"
version = "0.1.0"
description = ""
authors = ["Aabhiraj412 <aabhiraj412@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
undetected-chromedriver = "^3.5.5"
selenium = "^4.26.1"
redis = "^5.2.0"
pandas = "^2.2.3"
mysql-connector-python = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
